Move the test suite about running command line to another single module.
